refactor(errors): move einval
